2010-02-04 21:01 UTC+0100 Viktor Szakats (harbour.01 syenar.hu)   * src/rtl/console.c   * src/rtl/shadow.c     ! Fixed three remaining places where color value was cast to BYTE type.     ; QUESTION1: What to do with BYTE in hb_fs*Drv() calls?                  Should it be int or unsigned int?     ; QUESTION2: What to do with BYTE types in checksum calculations and                  low level color storage, is HB_U8 type good?
  * src/rtl/transfrm.c
  * src/rtl/memofile.c
    * BYTE -> char

  * src/rtl/gtkbstat.c
    * Formatting.

+++ harbour/src/rtl/console.c
@@ -641,7 +641,7 @@ HB_FUNC( HB_DISPOUTATBOX )
       iColor = hb_gtGetCurrColor();
 
    while( nStringLen-- )
-      hb_gtPutChar( iRow, iCol++, ( BYTE ) iColor, HB_GT_ATTR_BOX, ( unsigned char ) *pszString++ );
+      hb_gtPutChar( iRow, iCol++, iColor, HB_GT_ATTR_BOX, ( unsigned char ) *pszString++ );
 }
 
 HB_FUNC( HB_GETSTDIN ) /* Return handle for STDIN */
